now?? also added galaxy test that passes in CPU again for coverage
