+++ config/model/basic_sleep_galaxy.yaml
@@ -1,0 +1,34 @@
+# @package _group_
+name: SleepPhase
+image:
+  slen: 30
+  n_bands: 1
+  background:
+    - 686.0
+galaxy:
+  slen: 51
+  latent_dim: 8
+decoder:
+  params:
+    n_bands: ${model.image.n_bands}
+    slen: ${model.image.slen}
+    tile_slen: ${model.image.slen}
+    ptile_padding: 0
+    prob_galaxy: 1.0
+    n_galaxy_params: ${model.galaxy.latent_dim}
+    max_sources: 1
+    mean_sources: 0.99
+    min_sources: 0
+    gal_slen: ${model.galaxy.slen}
+    decoder_file: ${paths.root}/data/galaxy_decoder_1_band.dat
+    psf_params_file: ${paths.root}/data/fitted_powerlaw_psf_params.npy
+    background_values: ${model.image.background}
+encoder:
+  params:
+    n_bands: ${model.image.n_bands}
+    slen: ${model.image.slen}
+    tile_slen: ${model.image.slen}
+    ptile_slen: ${model.image.slen}
+    max_detections: ${model.decoder.params.max_sources}
+    n_galaxy_params: ${model.galaxy.latent_dim}
+    background_pad_value: ${model.image.background.0}
